+++ ext/duckdb/ruby-duckdb.h
@@ -30,6 +30,7 @@
 #include "./instance_cache.h"
 #include "./value_impl.h"
 #include "./scalar_function.h"
+#include "./scalar_function_set.h"
 #include "./expression.h"
 #include "./client_context.h"
 #include "./scalar_function_bind_info.h"